refactoring

+++ src/main/java/forex/checker/M3Checker.java
@@ -14,7 +14,7 @@ import java.util.concurrent.Callable;
 public class M3Checker implements Callable<Boolean> {
     private List<Price> priceList;
 
-    private void resetSimpleGrid(SimpleGrid simpleGrid, float minGrid) {
+    private void resetSimpleGrid(SimpleGrid simpleGrid, int minGrid) {
         simpleGrid.setMinGrid(minGrid);
         simpleGrid.setPulseCount(1);
         simpleGrid.setRollbackCount(0);
