+++ src/main/java/forex/processing/GridGenerator.java
@@ -40,7 +40,7 @@ public class GridGenerator {
         }
     }
 
-    private void resetSimpleGrid(SimpleGrid simpleGrid, float minGrid) {
+    private void resetSimpleGrid(SimpleGrid simpleGrid, int minGrid) {
         simpleGrid.setMinGrid(minGrid);
         simpleGrid.setPulseCount(1);
         simpleGrid.setRollbackCount(0);
